feat: move scanner
