+++ include/scanner.h
@@ -1,6 +1,6 @@
 #ifndef _THOR_SCANNER_H_
 #define _THOR_SCANNER_H_
-
+#include "trie.h"
 typedef enum
 {
     // Single character tokens
@@ -73,5 +73,37 @@ typedef struct
 
 void init_scanner(const char *);
 Token scan_token();
+
+// This is crappy
+Trie trie;
+
+void boostrap_keyword()
+{
+#define ADD_KEYWORD(word, token_type)     \
+    insert_trie(&trie, word, token_type); \
+    init_trie(&trie);
+    ADD_KEYWORD("and", TOKEN_AND)
+    ADD_KEYWORD("class", TOKEN_CLASS)
+    ADD_KEYWORD("else", TOKEN_ELSE)
+    ADD_KEYWORD("if", TOKEN_IF)
+    ADD_KEYWORD("nil", TOKEN_NIL)
+    ADD_KEYWORD("or", TOKEN_OR)
+    ADD_KEYWORD("print", TOKEN_PRINT)
+    ADD_KEYWORD("return", TOKEN_RETURN)
+    ADD_KEYWORD("super", TOKEN_SUPER)
+    ADD_KEYWORD("while", TOKEN_WHILE)
+    ADD_KEYWORD("false", TOKEN_FALSE)
+    ADD_KEYWORD("for", TOKEN_FOR)
+    ADD_KEYWORD("fun", TOKEN_FUN)
+    ADD_KEYWORD("this", TOKEN_THIS)
+    ADD_KEYWORD("true", TOKEN_TRUE)
+    ADD_KEYWORD("var", TOKEN_VAR)
+#undef ADD_KEYWORD
+}
+
+void cleanup_keyword()
+{
+    free_trie(&trie);
+}
 
 #endif // _THOR_SCANNER_H_
